rpa-bots/requirements.txt
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
